Add exception handlers to graph
For now this only generates the graph according to the exception
blocks defined in the bytecode. It does not perform any sophisticated
checking yet - in fact it still ignores any exception handlers and
assumes that they do not occur. It does, however, add elements to the
graph that we can expand further.