Simplify try-catch handling
Instead of assuming we could fail at every step, which introduced a
complex CFG and many new states, we simply assume that everything is
going to work, but collect all the objects that are used in the
try-block. Since we're doing worst-case analysis anyways, or
assumption must be that any number of the objects used in the
try-block has been 'tainted' and should now be in an unknown state,
hence we can simply transfer all objects to their exception state.